requirements.txt
nltk==3.8.1
contractions
autocorrect
flask
flask-cors
PyPDF2==3.0.1
python-docx==1.1.0